[Snap 2654] Remove duplicate entries from oldEntryMap (#446)
* Fix for memory leak in oldEntryMap
  When no transactions are running, take a write lock on rvv and clear
the map
  In some cases, entry in oldEntryMap and in region could have same
entryVersion. In that case, do clear the entry from oldEntryMap.
  In some cases, specially after truncate the entry in region could have
lower entryVersion. In that case, find out if there are any transaction
which requires the entry in oldEntryMap and if not so then delete it
from oldEntryMap
* Return TOMBSTONE when no entry present in oldEntryMap
* If entry is present in region, then check if it is valid for all
running transactions. If it is, then remove the entry present in
oldEntryMap.

* Account for the memory while clearing oldEntryMap